chore: improve Toolti[ exports
